Moved all order files into their own folder
